requirements.txt
PySide2==5.13.2
PyYAML==6.0.2
shiboken2==5.13.2
